requirements.txt
babel==2.16.0
chardet==5.2.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
fpdf==1.7.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
reportlab==4.2.5
six==1.17.0
tkcalendar==1.6.1
